chen-study/backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
reportlab==4.0.7
Pillow==10.1.0
